+++ lib/flame/fly_backend.ex
@@ -1,6 +1,6 @@
 defmodule FLAME.FlyBackend do
   @moduledoc """
-  The `FLAME.Backend` using [Fly.io](https://fly.io) machines.
+  A `FLAME.Backend` using [Fly.io](https://fly.io) machines.
 
   The only required configuration is telling FLAME to use the
   `FLAME.FlyBackend` by default and the `:token` which is your Fly.io API
@@ -20,27 +20,66 @@ defmodule FLAME.FlyBackend do
   ```
 
   The following backend options are supported, and mirror the
-  (Fly.io machines create API)[https://fly.io/docs/machines/working-with-machines/]:
+  [Fly.io machines create API](https://fly.io/docs/machines/api/machines-resource/#machine-config-object-properties):
 
-    * `:cpu_kind` - The size of the runner CPU. Defaults to `"performance"`.
+  * `:cpu_kind` - The size of the runner CPU. Defaults to `"performance"`.
 
-    * `:cpus` - The number of runner CPUs. Defaults to  `System.schedulers_online()`
-      for the number of cores of the running parent app.
+  * `:gpu_kind` - The type of GPU reservation to make.
 
-    * `:memory_mb` - The memory of the runner. Must be a 1024 multiple. Defaults to `4096`.
+  * `:cpus` - The number of runner CPUs. Defaults to  `System.schedulers_online()`
+    for the number of cores of the running parent app.
 
-    * `:boot_timeout` - The boot timeout. Defaults to `30_000`.
+  * `:memory_mb` - The memory of the runner. Must be a 1024 multiple. Defaults to `4096`.
 
-    * `:app` – The name of the otp app. Defaults to `System.get_env("FLY_APP_NAME")`,
+  * `:boot_timeout` - The boot timeout. Defaults to `30_000`.
 
-    * `:image` – The URL of the docker image to pass to the machines create endpoint.
-      Defaults to `System.get_env("FLY_IMAGE_REF")` which is the image of your running app.
+  * `:app` – The name of the otp app. Defaults to `System.get_env("FLY_APP_NAME")`,
 
-    * `:token` – The Fly API token. Defaults to `System.get_env("FLY_API_TOKEN")`.
+  * `:image` – The URL of the docker image to pass to the machines create endpoint.
+    Defaults to `System.get_env("FLY_IMAGE_REF")` which is the image of your running app.
 
-    * `:host` – The host of the Fly API. Defaults to `"https://api.machines.dev"`.
+  * `:token` – The Fly API token. Defaults to `System.get_env("FLY_API_TOKEN")`.
 
-    * `:services` - The optional services to run on the machine. Defaults to `[]`.
+  * `:host` – The host of the Fly API. Defaults to `"https://api.machines.dev"`.
+
+  * `:init` – The init object to pass to the machines create endpoint. Defaults to `%{}`.
+    Possible values include:
+
+      * `:cmd` – list of strings for the command
+      * `:entrypoint` – list strings for the entrypoint command
+      * `:exec` – list of strings for the exec command
+      * `:kernel_args` - list of strings
+      * `:swap_size_mb` – integer value in megabytes for th swap size
+      * `:tty` – boolean
+
+  * `:services` - The optional services to run on the machine. Defaults to `[]`.
+
+  * `:metadata` - The optional map of metadata to set for the machine. Defaults to `%{}`.
+
+  ## Environment Variables
+
+  The FLAME Fly machines do *do not* inherit the environment variables of the parent.
+  You must explicit provide the environment that you would like to forward to the
+  machine. For example, if your FLAME's are starting your Ecto repos, you can copy
+  the env from the parent:
+
+  ```elixir
+  config :flame, FLAME.FlyBackend,
+    token: System.fetch_env!("FLY_API_TOKEN"),
+    env: %{
+      "DATABASE_URL" => System.fetch_env!("DATABASE_URL")
+      "POOL_SIZE" => "1"
+    }
+  ```
+
+  Or pass the env to each pool:
+
+  ```elixir
+  {FLAME.Pool,
+    name: MyRunner,
+    backend: {FLAME.FlyBackend, env: %{"DATABASE_URL" => System.fetch_env!("DATABASE_URL")}}
+  }
+  ```
   """
   @behaviour FLAME.Backend
 
@@ -51,6 +90,7 @@ defmodule FLAME.FlyBackend do
   @derive {Inspect,
            only: [
              :host,
+             :init,
              :cpu_kind,
              :cpus,
              :gpu_kind,
@@ -61,15 +101,17 @@ defmodule FLAME.FlyBackend do
              :runner_id,
              :local_ip,
              :remote_terminator_pid,
-             :runner_node_basename,
              :runner_instance_id,
              :runner_private_ip,
+             :runner_node_base,
              :runner_node_name,
              :boot_timeout
            ]}
   defstruct host: nil,
+            init: %{},
             local_ip: nil,
             env: %{},
+            region: nil,
             cpu_kind: nil,
             cpus: nil,
             memory_mb: nil,
@@ -77,26 +119,29 @@ defmodule FLAME.FlyBackend do
             mounts: [],
             image: nil,
             services: [],
+            metadata: %{},
             app: nil,
             token: nil,
             boot_timeout: nil,
             runner_id: nil,
             remote_terminator_pid: nil,
             parent_ref: nil,
-            runner_node_basename: nil,
             runner_instance_id: nil,
             runner_private_ip: nil,
-            runner_node_name: nil
+            runner_node_base: nil,
+            runner_node_name: nil,
+            log: nil
 
   @valid_opts ~w(app image token host cpu_kind cpus mounts memory_mb gpu_kind boot_timeout env terminator_sup log services)a
 
   @impl true
   def init(opts) do
     conf = Application.get_env(:flame, __MODULE__) || []
-    [node_base | ip] = node() |> to_string() |> String.split("@")
+    [_node_base, ip] = node() |> to_string() |> String.split("@")
 
     default = %FlyBackend{
       app: System.get_env("FLY_APP_NAME"),
+      region: System.get_env("FLY_REGION"),
       image: System.get_env("FLY_IMAGE_REF"),
       token: System.get_env("FLY_API_TOKEN"),
       host: "https://api.machines.dev",
@@ -106,7 +151,10 @@ defmodule FLAME.FlyBackend do
       boot_timeout: 30_000,
       runner_node_basename: node_base,
       services: [],
-      mounts: []
+      mounts: [],
+      metadata: %{},
+      init: %{},
+      log: Keyword.get(conf, :log, false)
     }
 
     provided_opts =
@@ -114,7 +162,7 @@ defmodule FLAME.FlyBackend do
       |> Keyword.merge(opts)
       |> Keyword.validate!(@valid_opts)
 
-    state = Map.merge(default, Map.new(provided_opts))
+    %FlyBackend{} = state = Map.merge(default, Map.new(provided_opts))
 
     for key <- [:token, :image, :host, :app] do
       unless Map.get(state, key) do
@@ -122,18 +170,24 @@ defmodule FLAME.FlyBackend do
       end
     end
 
+    state = %FlyBackend{state | runner_node_base: "#{state.app}-flame-#{rand_id(20)}"}
     parent_ref = make_ref()
 
     encoded_parent =
       parent_ref
-      |> FLAME.Parent.new(self(), __MODULE__)
+      |> FLAME.Parent.new(self(), __MODULE__, state.runner_node_base, "FLY_PRIVATE_IP")
       |> FLAME.Parent.encode()
 
     new_env =
-      Map.merge(
-        %{PHX_SERVER: "false", FLAME_PARENT: encoded_parent},
-        state.env
-      )
+      %{"PHX_SERVER" => "false", "FLAME_PARENT" => encoded_parent}
+      |> Map.merge(state.env)
+      |> then(fn env ->
+        if flags = System.get_env("ERL_AFLAGS") do
+          Map.put_new(env, "ERL_AFLAGS", flags)
+        else
+          env
+        end
+      end)
 
     new_state =
       %FlyBackend{state | env: new_env, parent_ref: parent_ref, local_ip: ip}
@@ -169,44 +223,51 @@ defmodule FLAME.FlyBackend do
     {result, div(micro, 1000)}
   end
 
-  defp get_volume_id(%FlyBackend{ mounts: [] }), do: {nil, 0}
-  defp get_volume_id(%FlyBackend{ mounts: mounts } = state) when is_list(mounts) do
-        {volumes, time} = get_volumes(state)
+  defp get_volume_id(%FlyBackend{mounts: []}), do: {nil, 0}
 
-        case volumes do
-          [] ->
-            {:error, "no volumes to mount"}
-          all_volumes ->
-            volume_ids_by_name = 
-              all_volumes
-              |> Enum.filter(fn vol ->
-                vol["attached_machine_id"] == nil
-                and vol["state"] == "created"
-              end)
-              |> Enum.group_by(&(&1["name"]), &(&1["id"]))
+  defp get_volume_id(%FlyBackend{mounts: mounts} = state) when is_list(mounts) do
+    {volumes, time} = get_volumes(state)
 
-              new_mounts = Enum.map_reduce(
-                mounts,
-                volume_ids_by_name,
-                fn mount, leftover_vols ->
-                  case List.wrap(leftover_vols[mount.name]) do
-                    [] ->
-                      raise ArgumentError, "not enough fly volumes with the name \"#{mount.name}\" to a FLAME child"
-                    [volume_id | rest] ->
-                      {%{mount | volume: volume_id}, %{leftover_vols | mount.name => rest}}
-                  end
-                end
-              )
+    case volumes do
+      [] ->
+        {:error, "no volumes to mount"}
 
-            {new_mounts, time}
-        end
+      all_volumes ->
+        volume_ids_by_name =
+          all_volumes
+          |> Enum.filter(fn vol ->
+            vol["attached_machine_id"] == nil and
+              vol["state"] == "created"
+          end)
+          |> Enum.group_by(& &1["name"], & &1["id"])
+
+        new_mounts =
+          Enum.map_reduce(
+            mounts,
+            volume_ids_by_name,
+            fn mount, leftover_vols ->
+              case List.wrap(leftover_vols[mount.name]) do
+                [] ->
+                  raise ArgumentError,
+                        "not enough fly volumes with the name \"#{mount.name}\" to a FLAME child"
+
+                [volume_id | rest] ->
+                  {%{mount | volume: volume_id}, %{leftover_vols | mount.name => rest}}
+              end
+            end
+          )
+
+        {new_mounts, time}
+    end
   end
+
   defp get_volume_id(_) do
     raise ArgumentError, "expected a list of mounts"
   end
 
   defp get_volumes(%FlyBackend{} = state) do
-      {vols, get_vols_time} = with_elapsed_ms(fn ->
+    {vols, get_vols_time} =
+      with_elapsed_ms(fn ->
         Req.get!("#{state.host}/v1/apps/#{state.app}/volumes",
           connect_options: [timeout: state.boot_timeout],
           retry: false,
@@ -214,12 +275,13 @@ defmodule FLAME.FlyBackend do
         )
       end)
 
-      {vols.body, get_vols_time}
+    {vols.body, get_vols_time}
   end
 
   @impl true
   def remote_boot(%FlyBackend{parent_ref: parent_ref} = state) do
     {mounts, volume_validate_time} = get_volume_id(state)
+
     {req, req_connect_time} =
       with_elapsed_ms(fn ->
         Req.post!("#{state.host}/v1/apps/#{state.app}/machines",
@@ -236,7 +298,7 @@ defmodule FLAME.FlyBackend do
                 cpu_kind: state.cpu_kind,
                 cpus: state.cpus,
                 memory_mb: state.memory_mb,
-                gpu_kind: state.gpu_kind,
+                gpu_kind: state.gpu_kind
               },
               auto_destroy: true,
               restart: %{policy: "no"},
@@ -249,15 +311,14 @@ defmodule FLAME.FlyBackend do
 
     remaining_connect_window = state.boot_timeout - req_connect_time - volume_validate_time
 
-    case req.body do
+    case resp do
       %{"id" => id, "instance_id" => instance_id, "private_ip" => ip} ->
         new_state =
           %FlyBackend{
             state
             | runner_id: id,
               runner_instance_id: instance_id,
-              runner_private_ip: ip,
-              runner_node_name: :"#{state.runner_node_basename}@#{ip}"
+              runner_private_ip: ip
           }
 
         remote_terminator_pid =
@@ -270,7 +331,12 @@ defmodule FLAME.FlyBackend do
               exit(:timeout)
           end
 
-        new_state = %FlyBackend{new_state | remote_terminator_pid: remote_terminator_pid}
+        new_state = %FlyBackend{
+          new_state
+          | remote_terminator_pid: remote_terminator_pid,
+            runner_node_name: node(remote_terminator_pid)
+        }
+
         {:ok, remote_terminator_pid, new_state}
 
       other ->
@@ -279,6 +345,84 @@ defmodule FLAME.FlyBackend do
   end
 
   defp rand_id(len) do
-    len |> :crypto.strong_rand_bytes() |> Base.encode64(padding: false) |> binary_part(0, len)
+    len
+    |> :crypto.strong_rand_bytes()
+    |> Base.encode16(case: :lower)
+    |> binary_part(0, len)
+  end
+
+  defp http_post!(url, opts) do
+    Keyword.validate!(opts, [:headers, :body, :connect_timeout, :content_type])
+
+    headers =
+      for {field, val} <- Keyword.fetch!(opts, :headers),
+          do: {String.to_charlist(field), val}
+
+    body = Keyword.fetch!(opts, :body)
+    connect_timeout = Keyword.fetch!(opts, :connect_timeout)
+    content_type = Keyword.fetch!(opts, :content_type)
+
+    http_opts = [
+      ssl:
+        [
+          verify: :verify_peer,
+          depth: 2,
+          customize_hostname_check: [
+            match_fun: :public_key.pkix_verify_hostname_match_fun(:https)
+          ]
+        ] ++ cacerts_options(),
+      connect_timeout: connect_timeout
+    ]
+
+    case :httpc.request(:post, {url, headers, ~c"#{content_type}", body}, http_opts, []) do
+      {:ok, {{_, 200, _}, _, response_body}} ->
+        Jason.decode!(response_body)
+
+      {:ok, {{_, status, reason}, _, resp_body}} ->
+        raise "failed POST #{url} with #{inspect(status)} (#{inspect(reason)}): #{inspect(resp_body)} #{inspect(headers)}"
+
+      {:error, reason} ->
+        raise "failed POST #{url} with #{inspect(reason)} #{inspect(headers)}"
+    end
+  end
+
+  defp cacerts_options do
+    cond do
+      certs = otp_cacerts() ->
+        [cacerts: certs]
+
+      Application.spec(:castore, :vsn) ->
+        [cacertfile: Application.app_dir(:castore, "priv/cacerts.pem")]
+
+      true ->
+        IO.warn("""
+        No certificate trust store was found.
+
+        A certificate trust store is required in
+        order to download locales for your configuration.
+        Since elixir_make could not detect a system
+        installed certificate trust store one of the
+        following actions may be taken:
+
+        1. Use OTP 25+ on an OS that has built-in certificate
+           trust store.
+
+        2. Install the hex package `castore`. It will
+           be automatically detected after recompilation.
+
+        """)
+
+        []
+    end
+  end
+
+  if System.otp_release() >= "25" do
+    defp otp_cacerts do
+      :public_key.cacerts_get()
+    rescue
+      _ -> nil
+    end
+  else
+    defp otp_cacerts, do: nil
   end
 end