+++ lib/flame/pool/supervisor.ex
@@ -1,8 +1,13 @@
 defmodule FLAME.Pool.Supervisor do
+  @moduledoc false
+
   use Supervisor
 
   def start_link(opts) do
-    Supervisor.start_link(__MODULE__, opts, name: __MODULE__)
+    name = Keyword.fetch!(opts, :name)
+    pool_sup = Module.concat(name, "PoolSup")
+
+    Supervisor.start_link(__MODULE__, opts, name: pool_sup)
   end
 
   def init(opts) do