+++ test/runner_test.exs
@@ -4,6 +4,7 @@ defmodule FLAME.RunnerTest do
   import Mox
 
   alias FLAME.{Runner, MockBackend}
+  alias FLAME.Test.CodeSyncMock
 
   # Make sure mocks are verified when the test exits
   setup :set_mox_global
@@ -23,7 +24,7 @@ defmodule FLAME.RunnerTest do
   }
 
   defp remote_boot(state) do
-    parent = FLAME.Parent.new(make_ref(), self(), MockBackend)
+    parent = FLAME.Parent.new(make_ref(), self(), MockBackend, "app-flame-1", "MY_HOST_IP")
     name = Module.concat(FLAME.TerminatorTest, to_string(System.unique_integer([:positive])))
     opts = [name: name, parent: parent]
     spec = Supervisor.child_spec({FLAME.Terminator, opts}, restart: :temporary)
@@ -103,7 +104,7 @@ defmodule FLAME.RunnerTest do
       )
 
     assert Runner.remote_boot(runner) == :ok
-    assert Runner.call(runner, fn -> :works end) == :works
+    assert Runner.call(runner, self(), fn -> :works end) == :works
     assert_receive :stopped
   end
 
@@ -112,9 +113,9 @@ defmodule FLAME.RunnerTest do
 
     assert Runner.remote_boot(runner) == :ok
     assert Runner.remote_boot(runner) == {:error, :already_booted}
-    assert Runner.call(runner, fn -> :works end) == :works
+    assert Runner.call(runner, self(), fn -> :works end) == :works
     refute_receive :stopped
-    assert Runner.call(runner, fn -> :still_works end) == :still_works
+    assert Runner.call(runner, self(), fn -> :still_works end) == :still_works
     ref = Process.monitor(runner)
     assert Runner.shutdown(runner) == :ok
     assert_receive :stopped
@@ -153,7 +154,7 @@ defmodule FLAME.RunnerTest do
     test "single use" do
       {:ok, runner} = mock_successful_runner(3, single_use: true)
       assert Runner.remote_boot(runner) == :ok
-      error = wrap_exit(runner, fn -> Runner.call(runner, fn -> raise "boom" end) end)
+      error = wrap_exit(runner, fn -> Runner.call(runner, self(), fn -> raise "boom" end) end)
       assert {:exit, {%RuntimeError{message: "boom"}, _}} = error
       assert_receive :stopped
       assert Runner.shutdown(runner) == :ok
@@ -164,11 +165,11 @@ defmodule FLAME.RunnerTest do
       Process.monitor(runner)
       assert Runner.remote_boot(runner) == :ok
 
-      error = wrap_exit(runner, fn -> Runner.call(runner, fn -> raise "boom" end) end)
+      error = wrap_exit(runner, fn -> Runner.call(runner, self(), fn -> raise "boom" end) end)
       assert {:exit, {%RuntimeError{message: "boom"}, _}} = error
       refute_receive :stopped
       refute_receive {:DOWN, _ref, :process, ^runner, _}
-      assert Runner.call(runner, fn -> :works end) == :works
+      assert Runner.call(runner, self(), fn -> :works end) == :works
       assert Runner.shutdown(runner) == :ok
     end
   end
@@ -182,7 +183,9 @@ defmodule FLAME.RunnerTest do
       assert Runner.remote_boot(runner) == :ok
 
       error =
-        wrap_exit(runner, fn -> Runner.call(runner, fn -> Process.sleep(timeout * 2) end) end)
+        wrap_exit(runner, fn ->
+          Runner.call(runner, self(), fn -> Process.sleep(timeout * 2) end)
+        end)
 
       assert error == {:exit, :timeout}
 
@@ -199,13 +202,15 @@ defmodule FLAME.RunnerTest do
       assert Runner.remote_boot(runner) == :ok
 
       error =
-        wrap_exit(runner, fn -> Runner.call(runner, fn -> Process.sleep(timeout * 2) end) end)
+        wrap_exit(runner, fn ->
+          Runner.call(runner, self(), fn -> Process.sleep(timeout * 2) end)
+        end)
 
       assert error == {:exit, :timeout}
 
       refute_receive :stopped
       refute_receive {:DOWN, _ref, :process, ^runner, _}
-      assert Runner.call(runner, fn -> :works end, 1234) == :works
+      assert Runner.call(runner, self(), fn -> :works end, timeout: 1234) == :works
       assert Runner.shutdown(runner) == :ok
     end
   end
@@ -226,7 +231,7 @@ defmodule FLAME.RunnerTest do
       Process.unlink(runner)
       Process.monitor(runner)
       assert Runner.remote_boot(runner) == :ok
-      assert Runner.call(runner, fn -> :works end) == :works
+      assert Runner.call(runner, self(), fn -> :works end) == :works
       assert_receive :stopped, timeout * 2
       assert_receive {:DOWN, _ref, :process, ^runner, _}
     end
@@ -245,6 +250,34 @@ defmodule FLAME.RunnerTest do
       Agent.update(agent, fn _ -> true end)
       assert_receive :stopped, timeout * 2
       assert_receive {:DOWN, _ref, :process, ^runner, _}
+    end
+  end
+
+  describe "code_sync" do
+    test "copy_paths: true, copies the code paths and extracts on boot" do
+      mock = CodeSyncMock.new()
+      # the 4th invocation is the rpc to diff code paths
+      {:ok, runner} = mock_successful_runner(4, code_sync: mock.opts)
+
+      Process.monitor(runner)
+      assert Runner.remote_boot(runner) == :ok
+      assert Runner.call(runner, self(), fn -> :works end, timeout: 1234) == :works
+      assert Runner.shutdown(runner) == :ok
+      # called on remote boot
+      assert_receive {CodeSyncMock, {_mock, :extract}}
+
+      # called on :works call
+      assert_receive {CodeSyncMock, {_mock, :extract}}
+    end
+
+    test "noops by default" do
+      {:ok, runner} = mock_successful_runner(3)
+
+      Process.monitor(runner)
+      assert Runner.remote_boot(runner) == :ok
+      assert Runner.call(runner, self(), fn -> :works end, timeout: 1234) == :works
+      assert Runner.shutdown(runner) == :ok
+      refute_receive {CodeSyncMock, _}
     end
   end
 end